Auth - API (direct_hash)
